Flytter ut operasjoner på bigdecimal-tidsljener til egen fil
